Move FieldConfigFetcher to ocs-commons
