+++ ocs-commons/pom.xml
@@ -12,7 +12,7 @@
 	</parent>
 
 	<artifactId>ocs-commons</artifactId>
-	<version>0.4.0</version>
+	<version>0.5.0</version>
 	<packaging>jar</packaging>
 
 	<description>Library that contains common models and utility for the open-commerce-search stack</description>
