+++ pom.xml
@@ -44,7 +44,7 @@
 			<dependency>
 				<groupId>de.cxp.ocs</groupId>
 				<artifactId>ocs-commons</artifactId>
-				<version>0.4.0</version>
+				<version>0.5.0</version>
 			</dependency>
 			<dependency>
 				<groupId>de.cxp.ocs</groupId>
