+++ search-service/pom.xml
@@ -7,7 +7,7 @@
 	<parent>
 		<groupId>de.cxp.ocs</groupId>
 		<artifactId>ocs-parent</artifactId>
-		<version>0.8.5</version>
+		<version>0.10.0-SNAPSHOT</version>
 		<relativePath>../</relativePath>
 	</parent>
